+++ imports/api/users/userMethods.tests.js
@@ -1,64 +1,92 @@
-// /* eslint-env mocha */
-// /* eslint-disable func-names, prefer-arrow-callback */
-//
-// import { chai } from 'meteor/practicalmeteor:chai';
-// import { Meteor } from 'meteor/meteor';
-// import StubCollections from 'meteor/hwillson:stub-collections';
-//
-// import './userMethods';
-//
-// chai.should();
-//
-// if (Meteor.isServer) {
-//   describe('User Methods', () => {
-//     beforeEach(() => {
-//       StubCollections.stub(Branches);
-//     });
-//
-//     afterEach(() => {
-//       StubCollections.restore();
-//     });
-//
-//     describe('branch.new', () => {
-//       it('insert a new branch into the collection', () => {
-//         Branches.find({}).count().should.equal(0);
-//         Meteor.call('branch.new', { name: 'newBranch' });
-//         Branches.find({}).count().should.equal(1);
-//         Branches.findOne().name.should.equal('newBranch');
-//       });
-//
-//       it('returns the id of the new branch', () => {
-//         const id = Meteor.call('branch.new', { name: 'newBranch' });
-//         Branches.findOne()._id.should.equal(id);
-//       });
-//
-//       it('accepts a name property', () => {
-//         (() => Meteor.call('branch.new', { name: 'newBranch' }))
-//           .should
-//           .not
-//           .throw();
-//       });
-//
-//       it('strips out extraneous properties', () => {
-//         (() => Meteor.call('branch.new', { name: 'newBranch', prop: 'prop' }))
-//           .should
-//           .throw();
-//       });
-//     });
-//
-//     describe('branch.save', () => {
-//       let branchId;
-//       beforeEach(() => {
-//         branchId = Branches.insert({ name: 'a' });
-//       });
-//
-//       afterEach(() => {
-//         StubCollections.restore();
-//       });
-//       it('updates the branch name', () => {
-//         Meteor.call('branch.save', branchId, { name: 'b' });
-//         Branches.findOne(branchId).name.should.equal('b');
-//       });
-//     });
-//   });
-// }
+/* eslint-env mocha */
+/* eslint-disable func-names, prefer-arrow-callback */
+
+import { chai } from 'meteor/practicalmeteor:chai';
+import { Meteor } from 'meteor/meteor';
+
+import { Customers } from '../customers/customersCollection';
+
+import './userMethods';
+
+chai.should();
+
+if (Meteor.isServer) {
+  describe('User Methods', () => {
+    describe('users.favoriteCustomer', () => {
+      let userId;
+      let customerId;
+      beforeEach(() => {
+        Meteor.users.remove({});
+        userId = Meteor.users.insert({ profile: {} });
+        Customers.remove({});
+        customerId = Customers.insert({});
+      });
+
+      it('adds the customerId to the favorite customers array', () => {
+        Meteor.call('users.favoriteCustomer', userId, customerId);
+        Meteor.users.findOne(userId).profile.favoriteCustomers.length
+              .should.equal(1);
+        Meteor.users.findOne(userId).profile.favoriteCustomers[0]
+          .should.equal(customerId);
+      });
+
+      it('throws an error if the userId is invalid', function () {
+        this.timeout(10000);
+        (() => Meteor.call('users.favoriteCustomer', 'x', customerId))
+          .should
+          .throw(Error);
+      });
+
+      it('throws an error if the customerId is invalid', () => {
+        (() => Meteor.call('users.favoriteCustomer', userId, 'x'))
+          .should
+          .throw(Error);
+      });
+
+      it('throws an error if the customerId is already a favorite', () => {
+        Meteor.call('users.favoriteCustomer', userId, customerId);
+        (() => Meteor.call('users.favoriteCustomer', userId, customerId))
+          .should
+          .throw(Error);
+      });
+    });
+
+    describe('users.unfavoriteCustomer', () => {
+      let userId;
+      let customerId;
+      beforeEach(() => {
+        Customers.remove({});
+        customerId = Customers.insert({});
+        Meteor.users.remove({});
+        userId =
+          Meteor.users.insert({ profile: { favoriteCustomers: [customerId] } });
+      });
+
+      it('removes the customerId to the favorite customers array', () => {
+        Meteor.call('users.unfavoriteCustomer', userId, customerId);
+        Meteor.users.findOne(userId).profile.favoriteCustomers.length
+              .should.equal(0);
+      });
+
+      it('throws an error if the userId is invalid', function () {
+        this.timeout(10000);
+        (() => Meteor.call('users.favoriteCustomer', 'x', customerId))
+          .should
+          .throw(Error);
+      });
+
+      it('throws an error if the customerId is invalid', () => {
+        (() => Meteor.call('users.favoriteCustomer', userId, 'x'))
+          .should
+          .throw(Error);
+      });
+
+      it('throws an error if the customerId is not already a favorite', () => {
+        Meteor.call('users.unfavoriteCustomer', userId, customerId);
+        (() => Meteor.call('users.unfavoriteCustomer', userId, customerId))
+          .should
+          .throw(Error);
+      });
+    });
+  });
+}